+++ packages/adapters/wagmi/src/client.ts
@@ -282,44 +282,69 @@ export class WagmiAdapter implements ChainAdapter {
           ReturnType<(typeof UniversalProvider)['init']>
         >
 
-        const siweParams = await this.options?.siweConfig?.getMessageParams?.()
+        const clientId = await provider.client?.core?.crypto?.getClientId()
+        if (clientId) {
+          this.appKit?.setClientId(clientId)
+        }
 
-        const isSiweEnabled = this.options?.siweConfig?.options?.enabled
+        let chainId = Number(NetworkUtil.caipNetworkIdToNumber(this.appKit?.getCaipNetwork()?.id))
+        let address: string | undefined = undefined
+        let isSuccessful1CA = false
+
+        const isSiweEnabled = this.appKit?.getIsSiweEnabled()
         const isProviderSupported = typeof provider?.authenticate === 'function'
-        const isSiweParamsValid = siweParams && Object.keys(siweParams || {}).length > 0
-        const siweConfig = this.options?.siweConfig
+        const supports1ClickAuth = isSiweEnabled && isProviderSupported
 
-        if (isSiweEnabled && isProviderSupported && isSiweParamsValid && siweConfig) {
-          // @ts-expect-error - setting requested chains beforehand avoids wagmi auto disconnecting the session when `connect` is called because it things chains are stale
-          await connector.setRequestedChainsIds(siweParams.chains)
-
+        if (supports1ClickAuth) {
           const { SIWEController, getDidChainId, getDidAddress } = await import(
             '@reown/appkit-siwe'
           )
+          if (!SIWEController.state._client) {
+            return
+          }
 
-          const chains = this.caipNetworks
-            ?.filter(network => network.chainNamespace === 'eip155')
-            .map(chain => chain.id) as string[]
+          const siweParams = await SIWEController?.getMessageParams?.()
+          const isSiweParamsValid = siweParams && Object.keys(siweParams || {}).length > 0
 
+          if (!isSiweParamsValid) {
+            return
+          }
+
+          let reorderedChains = this.wagmiChains.map(chain => chain.id)
+
+          // @ts-expect-error - setting requested chains beforehand avoids wagmi auto disconnecting the session when `connect` is called because it thinks chains are stale
+          await connector.setRequestedChainsIds(reorderedChains)
+
+          if (chainId) {
+            reorderedChains = [chainId, ...reorderedChains.filter(c => c !== chainId)]
+          }
+
+          SIWEController.setIs1ClickAuthenticating(true)
           const result = await provider.authenticate({
-            nonce: await siweConfig.getNonce(),
+            nonce: await SIWEController.getNonce(),
             methods: [...OPTIONAL_METHODS],
             ...siweParams,
-            chains
+            chains: reorderedChains.map(chain => `eip155:${chain}`)
           })
+
           // Auths is an array of signed CACAO objects https://github.com/ChainAgnostic/CAIPs/blob/main/CAIPs/caip-74.md
           const signedCacao = result?.auths?.[0]
 
           if (signedCacao) {
             const { p, s } = signedCacao
             const cacaoChainId = getDidChainId(p.iss)
-            const address = getDidAddress(p.iss)
+            address = getDidAddress(p.iss)
+
             if (address && cacaoChainId) {
+              chainId = parseInt(cacaoChainId, 10)
+
               SIWEController.setSession({
                 address,
                 chainId: parseInt(cacaoChainId, 10)
               })
             }
+
+            SIWEController.setStatus('authenticating')
 
             try {
               // Kicks off verifyMessage and populates external states
@@ -331,22 +356,36 @@ export class WagmiAdapter implements ChainAdapter {
               await SIWEController.verifyMessage({
                 message,
                 signature: s.s,
-                cacao: signedCacao
+                cacao: signedCacao,
+                clientId
               })
+              isSuccessful1CA = true
             } catch (error) {
+              isSuccessful1CA = false
+              SIWEController.setIs1ClickAuthenticating(false)
+
               // eslint-disable-next-line no-console
               console.error('Error verifying message', error)
               // eslint-disable-next-line no-console
               await provider.disconnect().catch(console.error)
-              // eslint-disable-next-line no-console
-              await SIWEController.signOut().catch(console.error)
+              await this.connectionControllerClient?.disconnect().catch(console.error)
+              SIWEController.setStatus('error')
               throw error
             }
           }
+          SIWEController.setIs1ClickAuthenticating(false)
         }
 
-        const chainId = Number(NetworkUtil.caipNetworkIdToNumber(this.appKit?.getCaipNetwork()?.id))
         await connect(this.wagmiConfig, { connector, chainId })
+        const { SIWEController } = await import('@reown/appkit-siwe')
+        if (supports1ClickAuth && address && chainId && isSuccessful1CA) {
+          SIWEController.setStatus('authenticating')
+          await SIWEController.onSignIn?.({
+            address,
+            chainId
+          })
+          SIWEController.setStatus('success')
+        }
       },
       connectExternal: async ({ id, provider, info }) => {
         if (!this.wagmiConfig) {