+++ packages/cdn/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @reown/appkit-cdn
 
+## 1.0.6-6594e108a.0
+
+### Patch Changes
+
+- AppKit Auth
+- Updated dependencies
+  - @reown/appkit-polyfills@1.0.6-6594e108a.0
+  - @reown/appkit-ethers5@1.0.6-6594e108a.0
+  - @reown/appkit-ethers@1.0.6-6594e108a.0
+  - @reown/appkit-wagmi@1.0.6-6594e108a.0
+
 ## 1.0.5
 
 ### Patch Changes